pyproject.toml
[project]
name = "dragonpaw-bot"
version = "2.0.0"
description = "Dragonpaw's Discord Bot"
license = "Apache-2.0"
authors = [{ name = "Ash Arnold", email = "ash@dragonpaw.org" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "aiohttp>=3.11,<4",
    "emojis>=0.7,<1",
    "hikari[speedups]>=2.1,<3",
    "hikari-lightbulb>=2.3,<3",
    "palettable>=3.3,<4",
    "Pillow>=11,<12",
    "pydantic>=2,<3",
    "pyyaml>=6,<7",
    "python-dotenv>=1.0,<2",
    "safer>=4.12,<5",
    "setuptools>=75,<76",
    "uvloop>=0.21,<1",
]

[dependency-groups]
dev = [
    "ipython>=8.9,<9",
    "ty>=0.0.1a7",
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ty.environment]
python-version = "3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
extend-select = ["I", "PL"]

[tool.ruff.lint.per-file-ignores]
"dragonpaw_bot/plugins/subday/chart.py" = ["PLR0913"]
"tests/**" = ["PLR2004"]
